Fixing Windows VirtualBox vagrant templates for Packer 0.5.x After Packer 0.5.x was released, running `vagrant up` on VirtualBox produced the following error:
   ERROR vagrant: #<Vagrant::Errors::VBoxManageError: There was an error while executing `VBoxManage`, a CLI used by Vagrant
    for controlling VirtualBox. The command and stderr is shown below.

    Command: ["modifyvm", "0ae9250c-75b8-43ed-80d7-59e12f6d2c5a", "--macaddress1", "{{ .BaseMacAddress }}"]

    Stderr: VBoxManage: error: Invalid MAC address format

It turns out that with Packer 0.5.x the
`config.vm.base_mac = "{{.BaseMacAddress }}" is no longer required, and if
it is present you must run `vagrant up` twice to get the VirtualBox baseboxes
to work (though it has no adverse effect on VMware).

Though, in order to not break the VMware templates by omitting the
BaseMacAddress, you must supply the appropriate scsi adapter driver
for Packer 0.5.x.

This commit contains both fixes.